feat!(nmide-core): Added JSPS
The JSPS, (JavaScript Plugin System), feature
gives the functionality for plugin developers to
develop plugins in JavaScript. This not only
allows developers to develop in all languages
that can compiled to JavaScript, it also
facilitates usage of existing code to be used,
like web-editors, which will increase the end user
experience.

Changes:
- Removed unused files
- Msg are handled at the frontend before being sent to the backend
- Installation of plugins takes place in the backend
- Fixed ModelFold
- Implemented MsgListener
- Improved typing and error-messages in NmideClient
- Changed window.plugins from Array to Map