test(nmide-core): Created tests for RSC
Since the IDE is "zero-core", it is very easy to
mimic, so extensive tests can, and should be made.
But, to test the React-State-Change, a lot of
functionality can, (and has been), extracted to
pure functions.
Parametrized tests for installing, initializing,
and updating has been made. Along with primitive
tests for the utility functions used, ensuring no
logical errors are made. Future work should look
into the possibility of creating a similar
testing suite for JS-Plugins, to enhance the DX
for Plugin Developers.

JSP-TS, (JavaScript Plugin Testing Suite), should
be prioritized, since 1), it's easier to
prototype, and, 2), the first plugin TBD, is
going to be written in the JSP system.