refactor(core): Moved plugins to test folder
These plugins are not used, but could be used for testing purposes, as
they could provide a wider range of actual parameters in a parameterized
test. There is still work remaining for this to work, but instead of
deleting these plugins, as they are no longer needed, they can stay
here.