chore(ide-view): init commit
ide-view is a plugin for utilizing the ide-framework plugin. Any other
plugin can, optionally, specify a location they are to be put in, using
the class-attribute. If the html is to be placed in the "explorer",
simply add "location-explorer" to the outermost-parent. This works for
other "locations" aswell, but for the best experience, they should exist
before "view".