feat(core): Started on server capabilities
Refactored project to easier expose the server and IDE functionalities.
Work is still needed on the frontend side to ensure separation of
concerns. Currently, the bundled JS-code does not function on the
server, because it still uses Tauri code, especially in the client. This
can and will be fixed by setting the client as a field in the window,
ensuring plugins can override it. As well as separating Tauri-relevant
code to its own module, which will be swapped for the server-specific
module.