chore(ide): Init commit
This project contains just the necessary modules for the application,
and nothing else. The reason for creating this extra project, is to ease
showcasing of the features, ease eventual building of the files, and
minimize redundant/unused code, as JavaScript/TypeScript/PureScript does
not have as good support for conditional compilation as Rust.