refactor(core): de-Tauri-fied core
The core application is no longer Tauri-ed, meaning it can be used as a
standalone web-server. This is still WIP for the server, as no
functionality has been found, that is similar to Tauri's event system,
but when this is found, it will be possible to inject it right away.