chore: deleted unused files
The testing project was created with now deleted plugins in mind, and
was out-dated. A similar project should be created in the future, to
satisfy the need for E2E testing, especially if C-plugins are created.

This is currently not the case.

Similarly, the wrapper-project had out-dated code, and has most likely
fallen out of scope.