feat(core): Started on improving state management
Currently, if there is a collision between two states from plugins, this
is "managed" by dropping both states. This might not be the wanted
behavior for plugin developers, so this functionality should be exposed.
For example, it might be preferred, if the offending field is of type
list, that this list is merged.

This functionality is not fully implemented yet.