refactor(core): Rework state management
Plugins can now add functionality to handle state collisions.
For example, an Error-logging plugin, can allow collisions on it's own
field "error", so if another plugin wants to report an error, they can
do so by returning a state with the field "error", which, on a
collision, is just combined into a list.

Added documentation.