feat(core): Improved plugin installation
refactored installers to be `Task` instead of Promise, added an extra
promise that checks if all modules are installed every 50 ms, by
checking if `window.moduleCount` === 0.
Everytime a `module` _installs_ itself, by pushing a `Module`-object to
`window.plugins`, `window.moduleCount` is decremented. It should be
incremented everytime by the installer, on every _successful_
installation of a `module`.