feat(core-std-lib): Added mapping on Value type
Can now map on the values in the wrapper enum Value, making it easier to
work with. Instead of having to check if a Value is an i32, and then add
one to it, we can use `.map` and pass a closure, which is of the type:
`i32 -> i32`, which will result in `Result<Value, ValueMapperError>` if
we map on something that is not a number. This is done by implementing
the trait `ValueMapper` for the passed closure, meaning it is possible
to add more mappings in the future. The correct mapping is inferred by
the compiler, so we don't have to worry about filling the types in the
method: `map<F, T, R, G>`, as it can infer the correct types by simply
adding types to the closure.