feat!(core-std-lib): Refactored Event type
Event is now an enum, module developers can create dialog events, which
are only handled by the Core, not re-sent to other Modules, for the
dialog pop-up that Tauri supports.

PreExit and PostInit are also enums, along with the "normal" event.

Removed module_name field, so also changed the Core trait to reflect
this.